[plugin.video.travel] 3.0.2 (#1153)
