+++ plugin.video.travel/README.txt
@@ -3,6 +3,7 @@ plugin.video.travel
 
 XBMC Addon for Travel Network Video website
 
+Version 3.0.2 website change
 Version 3.0.1 Isengard version - separate scraper
 Version 1.0.4 website change
 Version 1.0.3 Added subtitles, metadata, views
