+++ plugin.video.travel/changelog.txt
@@ -1,3 +1,4 @@
+Version 3.0.2 website change
 Version 3.0.1 Isengard version - separate scraper
 Version 1.0.4 website change
 Version 1.0.3 Added subtitles, metadata, views
